everything can be deleted now

+++ customer_relations_manager/Controllers/StagesController.cs
@@ -72,6 +72,9 @@ namespace customer_relations_manager.Controllers
         [Authorize(Roles = nameof(UserRole.Super))]
         public void Delete(int id)
         {
+            var toDelete = _repo.GetByKey(id);
+            if (toDelete == null) return;
+            toDelete.ProductionViewSettings.Clear();
             _repo.DeleteByKey(id);
             _uow.Save();
         }
